+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* `print_bytes()` takes an `indent` argument
 
 ### 1.0.9
 * Raise `FileNotFoundError` instead of `ValueError` if provided YARA rules files or dirs don't exist