+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT RELEASE
 * `print_bytes()` takes an `indent` argument
+* Make `DecodingTableRow` and `BytesMatch` into dataclasses
 
 ### 1.0.9
 * Raise `FileNotFoundError` instead of `ValueError` if provided YARA rules files or dirs don't exist